Create a separate main file and clean up the code
The implementation of trilc should be independent of the IL generation
module. This commit therefore separates the two.

[ci skip] because no ci tests are implemented yet
